Move AssetIndexingSupport out of YAML namespace
